% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvSysSample.r
\name{EnvSysSample}
\alias{EnvSysSample}
\title{R function to perform a environmental systemic sampling design}
\usage{
EnvSysSample(envData,nExpect,plot=TRUE,saveShape=TRUE,nf,lowerLim,upperLim)
}
\arguments{
\item{envData}{May be a SpatialPointsDataFrame, a SpatilPixelsDataFrame, or a SpatialGridDataFrame as defined in package sp or a RasterStack or RasterBrick as defined in package raster. Note that this object can contain a mixed type variables (quantitative, factor and ordered)}

\item{nExpect}{Numeric Number of desired sampling points. Note that the exact expected number can not be reached because this function tries to find a grid that best matches the expected number}

\item{plot}{Logic TRUE  if you would like to have a graphical representation of the environmental systemic sampling design}

\item{nf}{Numeric of ordination axes retained in the analysis, this function manages currently only 2 or 3 dimentions}

\item{lowerLim}{Numeric value of probability in [0,1] that can be used to produce sample quantiles corresponding to the given lower limit probability. This may be used to reduce the effect of extremes observations}

\item{upperLim}{Numeric value of probability in [0,1] that can be used to produce sample quantiles corresponding to the given upper limit probability. This may be used to reduce the effect of extremes observations}

\item{saveShape=Logic}{TRUE if you want to save the geographical coordinates of the pixels corresponding to this systemic sampling design as a shapefile format. The generated shapefile will be saved in your working directory}
}
\value{
Returns a list containing 4 objects GeoSamples (The geographical coordinates of the pixels corresponding to the systemic sampling design).EnvSamples (The coordinates in the reduced environmental space of of the pixels corresponding to the systemic sampling design).EnvGrid (The coordinates of the grid in the reduced environmental).cellsize (he sizes of the grid cells in the different dimensions)
}
\description{
R function to perform a environmental systemic sampling design
}
\details{
The environmental systemic sampling design consists in reducing the environmental space in 2 or 3 dimensions using an ordination method (the dudi.mixt method in ade4 package) as a first step. Thereafter convex hull will be created around the reduced environmental space. Then, a non-regular multidimensional grid comprising will b ecreated inside the convex hull. The obtained grid represents the perfect configuration required to adequately survey the environmental space in the study area. The last step is subsequently to seek the closest pixel to this ideal configuration. This is achieved by searching the nearest neighbour between each grid point and each pixel in the environmental space based on Euclidean distances
}
\examples{
library(raster)
envData<-getData('worldclim', var='bio', res=10)
envData<-crop(envData,extent(-10,45,20,75))
plot(envData)
par(mfrow=c(1,2))
Mysampling1<-EnvSysSample(envData,nExpect=50,plot=T,saveShape=T,nf=2,lowerLim=0.00001,upperLim=0.99999)
plot(envData[[1]])
plot(Mysampling1[[1]],add=T,col=2,pch=19)
par(mfrow=c(1,2))
Mysampling2<-EnvSysSample(envData,nExpect=50,plot=T,saveShape=T,nf=2,lowerLim=0.1,upperLim=0.9)
plot(envData[[1]])
plot(Mysampling2[[1]],add=T,col=2,pch=19)
Mysampling3<-EnvSysSample(envData,nExpect=50,plot=T,saveShape=T,nf=3,lowerLim=0.001,upperLim=0.999)
plot(envData[[1]])
plot(Mysampling3[[1]],add=T,col=2)
}

