% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invasionForce.r
\name{invasionForce}
\alias{invasionForce}
\title{R function to compute a negative exponential dispersal kernel}
\usage{
invasionForce(occData,coords,a,envData,binary=TRUE,longlat=NULL)
}
\arguments{
\item{occData}{May be a SpatialPointsDataFrame as defined in package sp, a data.frame or  a matrix object containing species data}

\item{coords}{Optional 2 columns matrix containing the X and Y coordinates of occData(only consider if occData is a a data.frame or  a matrix object)}

\item{a}{An integer between 0 and 1 that controls the form of the dispersal kernel}

\item{envData}{May be a SpatialPointsDataFrame or a SpatialGridDataFrame as defined in package sp or a RasterStack or RasterBrick as defined in package raster. This object will be used to determine the area for which the negative exponential dispersal kernel will be calculated}

\item{binary}{logical TRUE if the occData object contains presences/absences data and FALSE if occData contains only the geographical coordinates of presences data}

\item{longlat}{Logical If TRUE, coordinates should be in degrees; else they should represent planar (Euclidean) space (e.g. units of meters)}
}
\value{
Returns a RasterLayer object of the negative exponential dispersal kernel
}
\description{
R function to compute a negative exponential dispersal kernel
}
\examples{
library(raster)
library(sp)
envData<-getData('worldclim', var='bio', res=10)
envData<-crop(envData,extent(-8,15,38,55))
envData<-envData[[1]]

#Generate randomly a SpatilaPointsDataFrame containing occurrences
xy<-coordinates(envData)[sample(which(is.na(values(envData))==FALSE),100),]
occ<-ifelse(xy[,2]>50,0,1)
occData<-SpatialPointsDataFrame(coords=xy,data=as.data.frame(occ), proj4string = CRS(proj4string(envData)))

par(mfrow=c(3,3),mar=c(1.5,1.5,1.5,1.5))
for (a in c(0.01,0.03,0.05,0.06,0.09,0.1,0.3,0.6,0.9)){
propagule<-invasionForce(occData,envData=envData,a=a,binary=TRUE,longlat=TRUE)
plot(propagule,main=paste("a = ",a))
plot(occData,col=ifelse(occData@data[,1]==1,1,0),add=TRUE,cex=0.3)}
}

