\name{iSDM-package}
\alias{iSDM-package}
\alias{iSDM}
\docType{package}
\title{
\packageTitle{iSDM}
}
\description{
\packageDescription{iSDM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{iSDM}
\packageIndices{iSDM}
This package contains the following 4 functions: EnvSysSample, invasionForce, likelihoodCA, RGBplot 
}
\author{
\packageAuthor{iSDM}

Maintainer: \packageMaintainer{iSDM}
}
\references{
Hattab et al xxxxxxxx (submited to Method in Ecology and Evolution)
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }


