% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGBplot.r
\name{RGBplot}
\alias{RGBplot}
\title{R function to plot the realized versus the potential distribution using a RGB colour space}
\usage{
RGBplot(x=x,y=y,cex,xlab=xlab,ylab=ylab)
}
\arguments{
\item{x}{May be a SpatilPixelsDataFrame, or a SpatialGridDataFrame as defined in package sp or a RasterLayer as defined in package raster}

\item{y}{An object belonging to the same class as x}

\item{cex}{A numerical value giving the amount by which plotting symbols should be magnified relative to the default}

\item{xlab}{A title for the x axis}

\item{ylab}{A title for the y axis}
}
\description{
R to plot the realized versus the potential distribution using a RGB colour space
}
\examples{
library(raster)
library(sp)
library(virtualspecies)
library(colorRamps)

envData<-getData('worldclim', var='bio', res=10)
envData<-crop(envData,extent(-8,15,38,55))

# Generate virtual species responses with formatfunctions
my.parameters <- formatFunctions(bio1 = c(fun = "dnorm", mean = 140, sd = 40), bio5 = c(fun = "dnorm",
mean = 230, sd = 70),bio6 = c(fun = "dnorm",mean = 10, sd = 40))
# Generate a virtual species potential distributions with responses to environmental variables
potential.dist <- generateSpFromFun(envData[[c(1,5,6)]], my.parameters)$suitab.raster

#Limit the distribution 
realized.dist<-potential.dist
cell.id<-which(coordinates(realized.dist)[,2]>48)
dis.lim<-sample(seq(0,1,by=0.01),length(cell.id),replace=TRUE)
values(realized.dist)[cell.id]<-ifelse(values(realized.dist)[cell.id]>dis.lim,
values(realized.dist)[cell.id]-dis.lim,0)
RGBplot(x=potential.dist,y=realized.dist,cex=0.6,xlab="Potential distribution"
,ylab="Realized distribution")
}
\references{
Tarek Hattab, Ruben Van De Kerchove, Ben Somers, Boris Brasseur, Carol Ximena Garz?n L?pez, Duccio Rocchini, Emilie Gallet-Moron, Fabien Spicher, Guillaume Decocq, Hannes Feilhauer, Hélène Horen, Jens Warrie, Michael Ewald, Olivier Honnay, Pieter Kempeneers, Raf Aerts, Sandra Skowronek, Sebastian Schmidtlein and Jonathan Lenoir (In prep).An unified framework to model the potential and realized distributions of invasive species within the invaded range
}

