\name{iSDM-package}
\alias{iSDM-package}
\alias{iSDM}
\docType{package}
\title{
\packageTitle{iSDM}
}
\description{
\packageDescription{iSDM}
}
\details{


}
\author{
\packageAuthor{iSDM}

Maintainer: \packageMaintainer{iSDM}
}
\references{
Tarek Hattab, Ruben Van De Kerchove, Ben Somers, Boris Brasseur, Carol Ximena Garzon Lopez, Duccio Rocchini, Emilie Gallet-Moron, Fabien Spicher, Guillaume Decocq, Hannes Feilhauer, Helene Horen, Jens Warrie, Michael Ewald, Olivier Honnay, Pieter Kempeneers, Raf Aerts, Sandra Skowronek, Sebastian Schmidtlein and Jonathan Lenoir (In prep). An unified framework to model the potential and realized distributions of invasive species within the invaded range
}

\keyword{ SDM, Species distribution modelling, invasive species,RGB plot }
\seealso{

}
\examples{

}
