\name{iSDM-package}
\alias{iSDM-package}
\alias{iSDM}
\docType{package}
\title{
\packageTitle{iSDM}
}
\description{
\packageDescription{iSDM}
}
\details{


}
\author{
\packageAuthor{iSDM}

Maintainer: \packageMaintainer{iSDM}
}
\references{
Tarek Hattab, Michael Ewald, Sandra Skowronek, Raf Aerts, Carol X. Garzón-López, Hélène Horen, Boris Brasseur, Emilie Gallet-Moron, Fabien Spicher, Guillaume Decocq, Hannes Feilhauer, Olivier Honnay, Pieter Kempeneers, Sebastian Schmidtlein, Ben Somers, Ruben Van De Kerchove, Duccio Rocchini and Jonathan Lenoir (In prep). An unified framework to model the potential and realized distributions of invasive species within the invaded range

}

\keyword{ SDM, species distribution modelling, invasive species, RGB plot }
\seealso{

}
\examples{

}
