% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodCA.r
\name{likelihoodCA}
\alias{likelihoodCA}
\title{R function to compute the likelihood of detecting contingent absences within a network of presence-absence data}
\usage{
likelihoodCA(occData,coords=NULL,envData, longlat=TRUE,nf=5)
}
\arguments{
\item{occData}{May be a SpatialPointsDataFrame as defined in package sp, a data.frame or  a matrix object containing species data (a single species) in binary format (ones for presences, zeros for absences)}

\item{coords}{Optional 2 columns matrix containing the X and Y coordinates of occData (only consider if occData is a a data.frame or  a matrix object)}

\item{envData}{An abject containing your explanatory variables. May be a SpatialPointsDataFrame or SpatialGridDataFrame as defined in package sp or a RasterStack or RasterBrick as defined in package raster. Note that this object can contain a mixed type variables (quantitative, factor and ordered).}

\item{longlat}{Logical TRUE if point coordinates are not projected(longitude-latitude decimal degrees)(+proj=longlat) and FALSE for plane (lonlat=FALSE)}

\item{nf}{If envData contains any factor and ordered variables, the Hill and Smith method will be used to perform an ordination of explanatory variables. nf corresponds in this case to the number of kept axes for the caluluation of mahalanobis distances.}
}
\value{
Returns a SpatialPointsDataFrame containing the likelihood values
}
\description{
R function to compute the likelihood of detecting contingent absences
}
\details{
This function campute the likelihood of detecting contingent absences within a network of presence-absence data. It's based on the idea that absences data that are environmentally close but geographically distant to presences data are more likely to be contingent absences.
}
\examples{
library(raster)
library(sp)
library(virtualspecies)
library(colorRamps)

envData<-getData('worldclim', var='bio', res=10)
envData<-crop(envData,extent(-8,15,38,55))

#Generate virtual species responses with formatfunctions 
my.parameters <- formatFunctions(bio1 = c(fun = "dnorm", mean = 140, sd = 40), bio5 = c(fun = "dnorm",
mean = 230, sd = 70),bio6 = c(fun = "dnorm",mean = 10, sd = 40))

#Generate a virtual species distributions with responses to environmental variables
potential.dist <- generateSpFromFun(envData[[c(1,5,6)]], my.parameters)

#Limit a virtual species distribution to a defined area. 
#It will thus generate a species which is not at the equilibrium with its environment
realized.dist<-limitDistribution(x=potential.dist$suitab.raster, area=extent(-8,15,38,48))

#Generate a random presence absence dataset from the realized distribution 
# using a probability threshold of 0.5
Presence<-coordinates(realized.dist$occupied.area
)[sample(which(values(realized.dist$occupied.area)>0.5),300),]
Absence<-coordinates(realized.dist$occupied.area
)[sample(which(values(realized.dist$occupied.area)<0.5),300),]
occData<-as.data.frame(rbind(cbind(Presence,SP=rep(1,300)),cbind(Absence,SP=rep(0,300))))
coordinates(occData)<-~x+y
proj4string(occData)<-proj4string(envData)

# Calculat  the likelihood of detecting contingent absences
likelihood<-likelihoodCA(occData=occData,envData=envData[[c(1,5,6)]],longlat=TRUE)

# Display the results
par(mfrow=c(1,2),mar=c(2,2.5,2,2.5))
plot(realized.dist$occupied.area,main="Realized distribution")
plot(occData,col=ifelse(occData$SP==1,2,1),add=TRUE,pch=19,cex=0.8)
plot(potential.dist$suitab.raster,main="Potential distribution")

scatterCol<-function(x){
x<-(x-min(x))/(max(x)-min(x))
colorFunction <- colorRamp(matlab.like(100))
zMatrix <- colorFunction(x)
zColors <- rgb(zMatrix[,1], zMatrix[,2], zMatrix[,3], maxColorValue=255)
return(zColors)}

points(likelihood,pch=21, col=1,bg=scatterCol(likelihood@data[,"LCA"]),cex=1)
}
\references{
Tarek Hattab, Ruben Van De Kerchove, Ben Somers, Boris Brasseur, Carol Ximena Garz?n L?pez, Duccio Rocchini, Emilie Gallet-Moron, Fabien Spicher, Guillaume Decocq, Hannes Feilhauer, Hélène Horen, Jens Warrie, Michael Ewald, Olivier Honnay, Pieter Kempeneers, Raf Aerts, Sandra Skowronek, Sebastian Schmidtlein and Jonathan Lenoir (In prep).An unified framework to model the potential and realized distributions of invasive species within the invaded range
}

